Small fixes

+++ src/main/resources/db/migration/V1__create_table.sql
@@ -4,14 +4,14 @@ CREATE TABLE cultural_building
     name             text NOT NULL,
     type_of_building text NOT NULL,
     address          text NOT NULL,
-    capacity         bigint NOT NULL,
+    capacity         integer NOT NULL,
     CONSTRAINT PK_cultural_buildings PRIMARY KEY ( id_place )
 );
 
 CREATE TABLE cinema
 (
     id_place    bigint NOT NULL,
-    screen_size bigint NOT NULL,
+    screen_size integer NOT NULL,
     screen_type text NOT NULL,
     CONSTRAINT PK_cinemas PRIMARY KEY ( id_place ),
     CONSTRAINT FK_24 FOREIGN KEY ( id_place ) REFERENCES cultural_building ( id_place )
@@ -21,7 +21,7 @@ CREATE TABLE theater
 (
     id_place            bigint NOT NULL,
     scene               text NOT NULL,
-    number_of_balconies bigint NOT NULL,
+    number_of_balconies integer NOT NULL,
     CONSTRAINT PK_theaters PRIMARY KEY ( id_place ),
     CONSTRAINT FK_21 FOREIGN KEY ( id_place ) REFERENCES cultural_building ( id_place )
 );
